tools/dev-testing/web-frontend/requirements.txt
flask>=2.0.0
psycopg2-binary>=2.9.0
pyyaml>=6.0
